Added the blockUser method.
Added the blockUser method, which blocks the OTHER user from view for THIS user. In this context, THIS user (i.e. the user logged in to the current device) is represented by uid_1, whereas the OTHER user (i.e. the user with whom this logged in user is interacting) is represented by uid_2. This method will NOT let THIS user block the OTHER user if the OTHER user is already blocking THIS user.