Added the addUserToEvent method and class.
Added the addUserToEvent method and class which allows an inviter to invite an invitee to an event. This method requires the eid (event identifier), inviter_uid (Inviter User Identifier), and invitee_uid (Invitee User Identifier). This information is then passed to the server, which fetches all possibly needed data to determine whether the invitee could practically be added to an event or not. The server then returns a message confirming success, explaining the failure, or showing an error.