Added a new method/class and updated an existing one.
Added a new method/class getImagesByPrivacyAndPurpose, which returns all of the images from the database table that are of the specified privacy and purpose. These images are in the form of Image objects which contain data, as seen in the Image class. Correspondingly, updated the Image class to allow direct access to attributes and removed get methods.

Modified the getEventDataByMember method/class to return not only the events to which the specified user belongs but also the events of which the specified user is a host.