Added two new FriendshipStatus methods/classes.
Added the getFriendshipStatusRequestSentUsers method/class, which returns a list of all users to whom the specified user has sent the specified request.
Added the getFriendshipStatusRequestReceivedUsers method/class, which returns a list of all users from whom the specified user has received the specified request.
Both of the above return a list of users, each of which has the following 4 fields populated: uid, username, firstName, lastName.

FROM NOW ON, THE ABOVE TWO METHODS SHOULD BE USED FOR ALL GET REQUESTS RELATED TO THE FRIENDSHIP STATUS.