Made various fixes. Added rejectFriend.
Added a class/method which takes two inputs: uid_1, which represents the user performing the action, and uid_2, which represents the user on whom the action is performed. This method will reject the friend request previously sent by uid_2 to uid_1. If such a request did not happen, a failure message will be returned instead.