Added the getEventDataByTopNLikes and getEventDataByTopNDislikes methods and subclasses to the EventRequest class.
The first Java method/subclass gets the data of an entire event for all of the events that have the top N count of likes, where N is the input value.
This second Java method/subclass gets the data of an entire event for all of the events that have the top N count of dislikes, where N is the input value.