Added the getEventDataByRadius class/method.
Added the getEventDataByRadius class/method to the EventRequest class in the ServerRequest package. If the longitude, latitude, and radius are provided, this method will return a JSON String which contains an array of Events and all of its associated data as well as the distance, in miles,  from the specified latitude and longitude.