Changed the incrementEventViewCount to incrementObjectViewCount.
Changed the incrementEventViewCount to incrementObjectViewCount. Instead of one parameter, two must now be supplied: the ID of the object, and the objectTypeLabel, which represents what object has been viewed. As of this moment, possible objects include "Event" and "User Image".
This method has been updated and moved to the MiscellaneousRequests java file.