Updated NewsFeedRequest Class and Methods.
Updated NewsFeedRequest Class and Methods to also return the username, first name, last name, and profile image data for the FriendUploadedImages news action.
Updated the Miscellaneous Class by added the convertPathsToFull method, which replaces local image paths provided by the server with full image paths that point to the file on that server.