Updated NewsFeedRequest Class and Methods.
Updated NewsFeedRequest Class and Methods to also return the profile image data for the FriendCreatedEvents news action.
Updated the Miscellaneous Class by fixing the convertPathsToFull method, which replaces local image paths provided by the server with full image paths that point to the file on that server and escaping slashes.