Updated the EventRequest.
Updated the EventRequest class and GetLiveEventDataByRadius to return the correct data (i.e. removed the ambiguity from the event host data). As a JSON string, not Java object.
Added the GetEventDataByRadius method, which was previously missing, to the EventRequest class. It is the same as the GetLiveEventDataByRadius method, except the data is not limited by the datetime.