Added and modified classes and methods.
Added the setEuiEventProfileSequence method to the EventUserImageRequest class, which sets the sequence of the event user image in the event profile.
Modified the UserImageRequest getImagesByEid class/method to allow for another input parameter. euiProfileIndicator, or Event User Image Profile Indicator narrows down the result set to event user images that belong to a profile (if input is set to "true"), event user images that do not belong to a profile (if input is set to "false"), or all event user images (if input is set to "null").
Updated Bryant's methods to handle the above change.