Updated the functionality of the UserImageRequest.setImage method.
Updated the functionality of the UserImageRequest.setImage method to now correctly handle setting values to null. An array of Booleans was introduced as an input parameter, in addition to the already existing parameters for this method. This array of Booleans must contain a quantity of Booleans equal to the amount of parameters this method accepts, not counting this array itself as well as the callback parameter. Before being sent to the backend server, it gets converted to a map of Booleans. Each Boolean represents whether or not this value is to be updated in the database.