Added the setEventUser method.
Added the setEventUser method to the EventUserRequest class, which updates the user with input uid of the event with input eid with the input values. The input values that are to update the current values may be actual values or NULL. Whether or not they are updated is determined by the setOrNot Boolean array, true representing that the value with the provided key should be updated.