Added, modified, and removed classes and methods.
Added the UserImageRequest.getImagesByUid subclass/method, which gets all of the images for the specified uid, optionally filtered on whether they are profile images or non-profile images. This subclass/method replaced UserImageRequest.getImagesByUidAndPurpose, as the "AndPurpose" part has been incorporated into the UserImageRequest.getImagesByUid through the userProfileIndicator parameter.
Modified the UserImageRequest.getImagesByEid subclass/method, to be as similar to getImagesByUid as possible. It now also returns the euiEventProfileSequence, euiInsertTimestamp, and euiUpdateTimestamp. (eui = Event User Image).
Removed the UserImageRequest.getImagesByPrivacyAndPurpose subclass/method. It was determined to be impractical, and is no needed in the presence of the two methods above.

Commented out Bryant's references to the methods that have been removed, and left the following tag: /** BRYANT UPDATE THIS **/