Added numerous tweaks and fixes.
Updated the FrienshipStatusRequest class with the getFriendshipStatusReceivedUsers and getFriendshipStatusSentUsers methods to (1) return as JSON and (2) return the userImagePath in addition to the previously returned data.
Added the SetOrNot Boolean to the UserRequest class setUser method to enable the functionality to modify a variable number of User properties at a time.